[REVIEW] - Use DataFusion optimizer source code instead of our own, refactor offset changes from upstream DataFusion
